Assignment #2 persistence of generated results

+++ src/main/java/tech/blackfall/dicedist/simulation/adapter/database/SimulationPartialResultEntity.java
@@ -15,5 +15,5 @@ class SimulationPartialResultEntity {
 
   private Long totalValue;
 
-  private Long occurrences;
+  private Integer occurrences;
 }
